Revert "feat: Short Url 상세 페이지 조회, Origin Url 로 리다이렉트하는 기능 구현."
This reverts commit f48b4dfb984378056e75dafe3ff8551a73321509.
