+++ src/main/java/com/programmers/springbooturlshortener/domain/url/UrlRepository.java
@@ -1,12 +1,10 @@
 package com.programmers.springbooturlshortener.domain.url;
 
-import java.util.Optional;
-
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
 
 public interface UrlRepository extends JpaRepository<Url, Long> {
 
     Optional<Url> findByOriginUrl(String originUrl);
-
-    Optional<Url> findByShortUrl(String shortUrl);
 }
