+++ src/main/java/com/programmers/springbooturlshortener/domain/url/dto/UrlCreateDto.java
@@ -1,21 +1,19 @@
 package com.programmers.springbooturlshortener.domain.url.dto;
 
+import com.programmers.springbooturlshortener.domain.url.Url;
+import org.hibernate.validator.constraints.URL;
+
 import javax.persistence.Column;
 import javax.validation.constraints.NotBlank;
 
-import org.hibernate.validator.constraints.URL;
-
 public record UrlCreateDto(@NotBlank @URL @Column(nullable = false) String originUrl,
-						   @NotBlank @Column(nullable = false) String algorithm) {
+                           @NotBlank @Column(nullable = false) String algorithm) {
 
-	public String removeProtocolFromOriginUrl() {
-		if (this.originUrl.startsWith("https://")) {
-			return this.originUrl.replace("https://", "");
-		}
-		if (this.originUrl.startsWith("http://")) {
-			return this.originUrl.replace("http://", "");
-		}
-
-		return this.originUrl;
-	}
+    public Url toEntity() {
+        return Url.builder()
+                .originUrl(originUrl)
+                .algorithm(algorithm)
+                .requestCount(1L)
+                .build();
+    }
 }
